bug(clustering): refactored that used old imports
