+++ docs/source/get_started.rst
@@ -128,7 +128,7 @@ Time Series Clustering
 .. code-block:: python
 
     >>> from sklearn.model_selection import train_test_split
-    >>> from sktime.clustering import TimeSeriesKMeans
+    >>> from sktime.clustering.k_means import TimeSeriesKMeans
     >>> from sktime.clustering.evaluation._plot_clustering import plot_cluster_algorithm
     >>> from sktime.datasets import load_arrow_head
