+++ sktime/clustering/k_medoids.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+"""Time series kmedoids."""
 __author__ = ["chrisholder", "TonyBagnall"]
 
 from typing import Callable, Union
